feat:  add root names to the textplan (#76)
Root names are added to both the parser and converter in this PR. The
root relation is not being treated as a relation internally and instead
is treated merely as annotation containing a list of names. This is
primarily because the relation-related codepaths make assumptions that
wouldn't apply to the root (such as having a valid Relation proto as its
data type).