fix:  support both old and new versions of protobuf
The version numbering scheme for the protobuf compiler was updated which causes a mismatch warning to appear for versions of cmake less than 3.27.  Later versions of protobuf are bundled with their own copy of abseil.  Since protobuf always required abseil anyway, this fix also updates the code to rely on that instead of our own submodule.  This should lead to greater consistency (especially with new protobuf code leaning on an experimental logging feature that won't be available in older versions of abseil).  If the newer CONFIG-based method of findpackage does not work, we now fall back to the older method of using findpackage.  The version of cmake that we require is still 3.20 as 3.27 isn't available in enough locations to advance to that minimum.