cleanup: Move install targets to CMake files under `export`.
Since the install logic needs to install everything that is part of the
public interface of the project, it is arguable best put in the CMake
files of the `export` folder, which also contains the public code. This
also removes the necessity to include the subdirectories under `src` in
a particular order, which would otherwise exist due to the `install`
command needing to see all installed targets.