fix: Preserve folder structure of installed proto header files.
The mechanism using the `PUBLIC_HEADERS` properties does not seem to
preserve the folder structure of the header files provided to it.
Consequently, `extensions/extensions.pb.h` ended up on the same level as
the other header files. This commit uses a different mechanism based on
`FILE_SET`, which, however, is only available since CMake v2.23.