feat: Make main target installable. (#103)
This allows to run `cmake --install .` to install the project to a
system such that it can be consumed with `find_package()`. This required
the following changes:

* Add `install(TARGET ...)` commands for all library and executable
targets.
* Add all public header files to the `HEADER` `FILE_SET` of their
respective target.
* Export all these targets to an export set and `install(EXPORT ...)`
that export set.
* Provide install-time paths for several `target_include_directories()`,
  some of which are empty because the headers are not installed.

Since some of the library targets depend on external projects
(`abseil-cpp`
and `yaml-cpp`), the install logic of these projects also needs to be
enabled.

This change is not only necessary to actually install the project, but
also to consume it via `add_subdirectory` from any project that,
itself, is installable. (All dependencies of installable targets must be
installable.)