Don't define member functions inside a nested namespace
This trips up MSVC, in that member functions aren't properly included in the object files of the various translation units. Generally, this should be fixed by having a `using` declaration in a .cpp file. From my experience, a general good practice is to never nest definitions in namespaces, unless that is only visible from within a translation unit (e.g. a `detail` namespace inside a .cpp file).

This doesn't fix all of the places in the code where this pattern is used - only the (narrow) compile path that i've been excercising. Someone should probably go through the rest of the codebase and apply this - will be needed if/when a windows CI pipeline is added.