Set ANTLR4CPP_STATIC for everything that links against ANTLR (#112)
It isn't sufficient to rely on a single
[`add_definitions`](https://github.com/substrait-io/substrait-cpp/blob/cc8d08af7a7ff4b65d0081fc18f9bb243fe85824/src/substrait/textplan/parser/grammar/CMakeLists.txt#L8)
when defining this. Other things which pulls in ANTLR headers _must_
define this as well - if not, we end up with very hard to debug symbol
visibility errors.

Set it as a public compile definition on the target to ensure the
definition propagates.