Don't throw away `visit` result in BasePlanProtoVisitor (#118)
`visitPlan` is marked `protected`, meaning that non-subclasses of
`BasePlanProtoVisitor` (which is probably what'll initiate the `visit`
most of the time) cannot have access to the `visit` result.

Is there a specific reason that the traversal result has been hidden? If
not, i suggest this change.

If so, subclasses are obviously free to implement a `visitWithResult`
function to do the same thing.