feat: update to Substrait 0.43, add min. precision timestamp support
Update to Substrait v0.43.0, which is the next version, and add minimal
support for precision timestamp support, which is the main change since
the previously used v0.42.1. The minimal support is achieved by adding
the new type cases to the type switches, where they either produce "not
implemented" errors or do the simple functionality of setting
nullability of the type, as well as suppressing warnings of using
depcreated protobuf accessors in the places where these were not
suppressed yet.

Signed-off-by: Ingo Müller <ingomueller@google.com>