fix: use proper string concatenation for error on unsupported types
The previous default template instances of `decodeType` for parametrized
types used a "concatenation" on two char pointers, which doesn't work.
This PR fixes that by making one of the two an `std::string`. The error
did not show up until now because all supported types have
specializations of the template.

Signed-off-by: Ingo Müller <ingomueller@google.com>