Windows build & test CI and fixes (#134)
`find_vs.ps1` is used to enter a visual studio development environment,
which in turn, makes `ninja` available.

This also includes a fix which i've been running with locally for a
while... For some yet unknown reason, the googletest dependency within
ANTLR4 (antlr4->googletest) applies incorrect assumptions about include
path inside the gmock project. This only happens when building with
MSVC. We can hot-fix this by injecting the correct paths to the target,
after it's been defined.

Unfortunately, there's a lingering issue
[here](https://github.com/mortbopet/substrait-cpp/blob/main/src/substrait/textplan/parser/SubstraitPlanSubqueryRelationVisitor.cpp)
where a std::any_cast exception occurs. I've spent a fair bit of time
trying to track down the issue, but to no avail. I don't plan on using
the textual format any time soon, so for the sake of progress, i've
disabled the text-format related tests for now. After merging this PR,
I'll create an issue to track it, in case some Windows user stumbles
upon this in the future.